+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
+## 2023.3.3
+
+fixed: force_screen skips imediatly to the selected screen
+added: hold_time configurable 
+
+## 2023.3.2
+
+added: hold_screen for 20 additional seconds
+
 ## 2023.3.1
 
 added: del_screen with wildcards
 changed: maximum icons to 80
 fixed: skip_next
+fixed: show_all_icons on boot
 
 ## 2023.3.0